Removes pause in playbook after droplet creation.
Since Ansible isn't smart enough to load group vars
for hosts that were added to the inventory via `add_host`,
the best way to provision a new host and then run a playbook
against it is to create a wrapper script that calls two playbooks
sequentially. Bah, but it works.